break out BootScriptSerializer
