+++ api/serializers/__init__.py
@@ -9,3 +9,4 @@ from .instance_related_field import InstanceRelatedField
 from .account_serializer import AccountSerializer
 from .provider_serializer import ProviderSerializer
 from .cleaned_identity_serializer import CleanedIdentitySerializer
+from .boot_script_serializer import BootScriptSerializer
